requirements.txt
llama-index
openai
python-dotenv
docx2txt
pypdf